Revert "Fix: カードのユーザー名取得処理でユーザー名が更新されないバグを直した"
